Exception handling - delete

+++ src/main/java/com/viniciuscunha/learningspringboot/services/UserService.java
@@ -2,8 +2,10 @@ package com.viniciuscunha.learningspringboot.services;
 
 import com.viniciuscunha.learningspringboot.entitites.User;
 import com.viniciuscunha.learningspringboot.repositories.UserRepository;
+import com.viniciuscunha.learningspringboot.services.exceptions.DataBaseException;
 import com.viniciuscunha.learningspringboot.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +31,14 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        try {
+            if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+            repository.deleteById(id);
+        } catch (ResourceNotFoundException e){
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
